requirements.txt
flask
flask_restful
requests
PyGithub
pythainlp
sklearn
paramiko
# jwt